% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_amt_for_dose}
\alias{check_amt_for_dose}
\title{AMT present for dosing records}
\usage{
check_amt_for_dose(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
AMT present for dosing records
}
