% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_bloq_middle}
\alias{check_bloq_middle}
\title{BLOQ in middle of profile}
\usage{
check_bloq_middle(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
BLOQ in middle of profile
}
