% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_char_num_mapping}
\alias{check_char_num_mapping}
\title{Check char-numeric mapping consistency (RACE vs RACEN)}
\usage{
check_char_num_mapping(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check char-numeric mapping consistency (RACE vs RACEN)
}
