% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_char_truncation}
\alias{check_char_truncation}
\title{Check potential character truncation}
\usage{
check_char_truncation(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check potential character truncation
}
