% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_covariate_outliers}
\alias{check_covariate_outliers}
\title{Check outliers in covariates}
\usage{
check_covariate_outliers(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check outliers in covariates
}
