% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Identify duplicate records}
\usage{
check_duplicates(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Identify duplicate records
}
