% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_event_ordering}
\alias{check_event_ordering}
\title{Check event ordering at identical subject-time points}
\usage{
check_event_ordering(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check event ordering at identical subject-time points
}
