% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_evid4_once}
\alias{check_evid4_once}
\title{EVID=4 appears once per period}
\usage{
check_evid4_once(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
EVID=4 appears once per period
}
