% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_expected_ranges}
\alias{check_expected_ranges}
\title{Check expected ranges for common fields}
\usage{
check_expected_ranges(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check expected ranges for common fields
}
