% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_fixed_covariates}
\alias{check_fixed_covariates}
\title{Check fixed covariates do not change within subject}
\usage{
check_fixed_covariates(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check fixed covariates do not change within subject
}
