% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_high_predose}
\alias{check_high_predose}
\title{Abnormally high predose concentrations}
\usage{
check_high_predose(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Abnormally high predose concentrations
}
