% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_mdv_assignment}
\alias{check_mdv_assignment}
\title{MDV assignment consistency}
\usage{
check_mdv_assignment(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
MDV assignment consistency
}
