% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_name_label_len}
\alias{check_name_label_len}
\title{Check variable name and label length constraints}
\usage{
check_name_label_len(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check variable name and label length constraints
}
