% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_nominal_actual_consistency}
\alias{check_nominal_actual_consistency}
\title{Check nominal and actual consistency (ordering)}
\usage{
check_nominal_actual_consistency(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check nominal and actual consistency (ordering)
}
