% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_nominal_actual_deviation}
\alias{check_nominal_actual_deviation}
\title{Check large nominal/actual deviation (absolute)}
\usage{
check_nominal_actual_deviation(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check large nominal/actual deviation (absolute)
}
