% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_obs_has_prior_dose}
\alias{check_obs_has_prior_dose}
\title{Check every observation has a prior dose within subject-period}
\usage{
check_obs_has_prior_dose(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check every observation has a prior dose within subject-period
}
