% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_pk_no_dose}
\alias{check_pk_no_dose}
\title{Check subjects with PK records but no dosing records}
\usage{
check_pk_no_dose(adppk, addose)
}
\arguments{
\item{adppk}{ADPPK data frame.}

\item{addose}{Dose-level data frame.}
}
\value{
List with check result.
}
\description{
Check subjects with PK records but no dosing records
}
