% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_predose_time}
\alias{check_predose_time}
\title{Check predose time <= 0}
\usage{
check_predose_time(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check predose time <= 0
}
