% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_required_vars}
\alias{check_required_vars}
\title{Check required ADPPK variables}
\usage{
check_required_vars(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check required ADPPK variables
}
