% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_standardized_values}
\alias{check_standardized_values}
\title{Standardize category and numeric attributes}
\usage{
check_standardized_values(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Standardize category and numeric attributes
}
