% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_time_anchor_consistency}
\alias{check_time_anchor_consistency}
\title{Check dose time anchor consistency by subject-period}
\usage{
check_time_anchor_consistency(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check dose time anchor consistency by subject-period
}
