% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_time_sequential}
\alias{check_time_sequential}
\title{Actual times increase sequentially after first dose}
\usage{
check_time_sequential(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Actual times increase sequentially after first dose
}
