% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_unexpected_values}
\alias{check_unexpected_values}
\title{Check unusual values for dose or concentration}
\usage{
check_unexpected_values(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check unusual values for dose or concentration
}
