% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{checks_to_issues}
\alias{checks_to_issues}
\title{Build issue table from check results}
\usage{
checks_to_issues(check_out)
}
\arguments{
\item{check_out}{Named list from \code{run_checks()}}
}
\value{
data.frame
}
\description{
Build issue table from check results
}
