% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{checks_to_summary}
\alias{checks_to_summary}
\title{Convert check results to summary data frame}
\usage{
checks_to_summary(check_out)
}
\arguments{
\item{check_out}{Named list from \code{run_checks()}}
}
\value{
data.frame
}
\description{
Convert check results to summary data frame
}
