% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_runner.R
\name{draw_readiness_gauge}
\alias{draw_readiness_gauge}
\title{draw readiness gauge}
\usage{
draw_readiness_gauge(m)
}
\arguments{
\item{m}{output from model_readiness_score}
}
\value{
ggplot object or base plot
}
\description{
draw readiness gauge
}
\keyword{internal}
