% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{generate_check_report_html}
\alias{generate_check_report_html}
\title{Generate HTML checklist report}
\usage{
generate_check_report_html(adppk, check_out, file, cfg = NULL)
}
\arguments{
\item{adppk}{ADPPK data frame}

\item{check_out}{Named list from \code{run_checks()}}

\item{file}{Output html path}

\item{cfg}{Optional check config list.}
}
\value{
Invisibly returns file path
}
\description{
Generate HTML checklist report
}
