% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{load_check_config}
\alias{load_check_config}
\title{Load check configuration from YAML}
\usage{
load_check_config(path = NULL)
}
\arguments{
\item{path}{Optional path to YAML file. If NULL, use built-in default config.}
}
\value{
A list containing check config.
}
\description{
Load check configuration from YAML
}
