% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_runner.R
\name{mod_check_runner}
\alias{mod_check_runner}
\title{teal module: check runner}
\usage{
mod_check_runner(id)
}
\arguments{
\item{id}{module id}
}
\value{
teal module
}
\description{
teal module: check runner
}
\seealso{
Other teal-modules: 
\code{\link{mod_data_overview}()},
\code{\link{mod_pk_visualization}()}
}
\concept{teal-modules}
