% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_overview.R
\name{mod_data_overview}
\alias{mod_data_overview}
\title{teal module: data overview}
\usage{
mod_data_overview(id)
}
\arguments{
\item{id}{module id}
}
\value{
teal module
}
\description{
Provides a browsable data preview, per-column missingness summary, and
dataset-level metrics for ADPPK.
}
\seealso{
Other teal-modules: 
\code{\link{mod_check_runner}()},
\code{\link{mod_pk_visualization}()}
}
\concept{teal-modules}
