% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pk_visualization.R
\name{mod_pk_visualization}
\alias{mod_pk_visualization}
\title{teal module: PK visualization}
\usage{
mod_pk_visualization(id)
}
\arguments{
\item{id}{module id}
}
\value{
teal module
}
\description{
teal module: PK visualization
}
\seealso{
Other teal-modules: 
\code{\link{mod_check_runner}()},
\code{\link{mod_data_overview}()}
}
\concept{teal-modules}
