% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_runner.R
\name{run_check_with_cfg}
\alias{run_check_with_cfg}
\title{Internal: apply check function with optional config item}
\usage{
run_check_with_cfg(fn, args, cfg_item = NULL)
}
\arguments{
\item{fn}{Check function}

\item{args}{Named list of args for function}

\item{cfg_item}{Optional config list for this check}
}
\value{
Check result list
}
\description{
Internal: apply check function with optional config item
}
