% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{run_checks}
\alias{run_checks}
\title{Run selected checks}
\usage{
run_checks(adppk, addose = data.frame(), selected = checks_registry()$id)
}
\arguments{
\item{adppk}{ADPPK data frame.}

\item{addose}{ADDOSE data frame.}

\item{selected}{Character vector of check ids.}
}
\value{
Named list of check results.
}
\description{
Run selected checks
}
