% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_cross_study_alignment}
\alias{check_cross_study_alignment}
\title{Align variable types/formats across studies}
\usage{
check_cross_study_alignment(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Returns a skip result for single-study datasets (check is not applicable).
}
