% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_high_predose}
\alias{check_high_predose}
\title{Abnormally high predose concentrations}
\usage{
check_high_predose(adppk, predose_ratio_threshold = 0.2)
}
\arguments{
\item{adppk}{ADPPK data frame.}

\item{predose_ratio_threshold}{Predose to subject max concentration ratio threshold.}
}
\value{
List with check result.
}
\description{
Abnormally high predose concentrations
}
