% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{check_sampling_deviation}
\alias{check_sampling_deviation}
\title{Check nominal vs actual time deviation >10\%}
\usage{
check_sampling_deviation(adppk)
}
\arguments{
\item{adppk}{ADPPK data frame.}
}
\value{
List with check result.
}
\description{
Check nominal vs actual time deviation >10\%
}
