% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_core.R
\name{run_checks}
\alias{run_checks}
\title{Run selected checks}
\usage{
run_checks(
  adppk,
  addose = data.frame(),
  selected = c("required_vars", "name_label_len", "pk_no_dose", "poppk_consistency",
    "char_num_mapping")
)
}
\arguments{
\item{adppk}{ADPPK data frame.}

\item{addose}{ADDOSE data frame.}

\item{selected}{character vector of check ids.}
}
\value{
List of check results.
}
\description{
Run selected checks
}
